Fixed delete bug that i introduced
git-svn-id: https://vimpc.svn.sourceforge.net/svnroot/vimpc/trunk@204 fc90b29e-3040-4360-8be2-b397c2834534

+++ src/window/browsewindow.cpp
@@ -169,7 +169,7 @@ uint32_t BrowseWindow::Current() const
    uint32_t current       = 0;
    int32_t  currentSongId = client_.GetCurrentSong();
 
-   if ((currentSongId >= 0) && (currentSongId < Main::Playlist().Size()))
+   if ((currentSongId >= 0) && (currentSongId < static_cast<int32_t>(Main::Playlist().Size())))
    {
       current = Main::Browse().Index(Main::Playlist().Get(currentSongId));
    }
